Some code tidy
